Basis of graph support
