+++ phpunit.xml.dist
@@ -17,7 +17,7 @@
     
     <testsuites>
         <testsuite name="BasicShopifyAPI Test Suite">
-            <file>./test/test.php</file>
+            <directory>./test/</directory>
         </testsuite>
     </testsuites>
