+++ test/GraphApiTest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace OhMyBrew\ShopifyAPI;
+
+use GuzzleHttp\Client;
+use GuzzleHttp\Handler\MockHandler;
+use GuzzleHttp\Psr7\Response;
+use ReflectionClass;
+
+class GraphApiTest extends \PHPUnit\Framework\TestCase
+{
+    /**
+     * @test
+     *
+     * Checking base URL for API calls on public
+     */
+    public function itShouldReturnBaseUrl()
+    {
+        $api = new GraphAPI();
+        $api->setShop('example.myshopify.com');
+
+        $this->assertEquals('https://example.myshopify.com/admin/api/graphql.json', $api->getBaseUrl());
+    }
+}
