+++ test/fixtures/rest/admin__shop.json
